ci: Test missing op definitions against `pytket` (#36)
Adds an integration tests that checks whether all elements of
`pytket.OpType.__members__` are present in the `OpType` enum.

Adds a CI check for this that runs on pull requests / merges as well as
on a cron job every day to ensure that new pytket releases are always
tested.

This test currently fails with the following message:
```
Missing optypes in `tket_json_rs`:
  - Phase
  - CS
  - CSdg
  - TermSequenceBox
  - ConjugationBox
  - DummyBox
  - CnY
  - CnZ
  - WASM
  - MultiplexedTensoredU2Box
  - StatePreparationBox
  - DiagonalBox
Please add them to the `OpType` enum in `tket_json_rs/src/optype.rs`.
```

I'll open another PR with those implementations.

Ideally the cron job would also create an issue when the test fails, but
I'll look into that another time.

Closes #31 .